+++ src/components/ContentTab/EntryContent/EntryContent.vue
@@ -739,7 +739,7 @@
   }
   
   .tags-container {
-    // TODO - search for "31" and see todo note about changing this to rem
+    // TODO - search for "31" and see toDo note about changing this to rem
     min-height: 43px; /* Set a fixed minimum height for the tags container */
     position: relative;
   }